% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Airflow_kernel.R
\name{f_airflow}
\alias{f_airflow}
\title{Function for the terrestrial airstream kernel.}
\usage{
f_airflow(D, b, a)
}
\arguments{
\item{D}{numeric, a vector of distances between the points (in Km)}

\item{b}{numeric, shape parameter. Should be always >0.}

\item{a}{numeric, scale parameter.}
}
\value{
numeric vector, including the probabilities of a species travelling the input distances by vehicle airstream.
}
\description{
This function computes the probability of a species to reach a point in the network
starting from a different point by vehicle airstream. Mathematically described as lognormal function (see Von der Lippe et al. 2013).
}
\examples{
#generating vector of distances, in Km:
dist<-seq(from=0.1,to=10,length.out=100)

#assigning parameter values:
par_a <- -0.724 # Von Der Lippe et al. 2013. Values for A. altissima
par_b <- 2.307 # Von Der Lippe et al. 2013. Values for A. altissima

#Calculating and plotting probabilities
AirstreamDisp<-f_airflow(D = dist, b = par_b, a = par_a)
plot(AirstreamDisp~D,xlab="Distance (Km)", ylab="Probability", pch=16)

}
\references{
von der Lippe, M., Bullock, J. M., Kowarik, I., Knopp, T., & Wichmann, M. (2013). Human-mediated dispersal of seeds by the airflow of vehicles. PloS one, 8(1), e52733.
}
\author{
Maurizio Bagnara, Hanno Seebens
}
\keyword{CASPIAN}
