% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Attachment_kernel.R
\name{f_attach}
\alias{f_attach}
\title{Function for the terrestrial attachment kernel.}
\usage{
f_attach(D, a, b, g)
}
\arguments{
\item{D}{numeric, a vector of distances between the points (in Km)}

\item{a}{numeric, shape parameter.}

\item{b}{numeric, scale parameter.}

\item{g}{numeric, scale parameter.}
}
\value{
numeric vector, including the probabilities of a species travelling the input distances by attachment to vehicles.
}
\description{
This function computes the probability of a species to reach a point in the network
starting from a different point by attachment to vehicles. Mathematically described as exponential function (see Taylor et al. 2012, model 2).
}
\examples{
#generating vector of distances, in Km:
dist<-seq(from=0.1,to=10,length.out=100)

#assigning parameter values:
par_a <- 0.6556  # parameter c in Taylor et al. 2012, Model 2
par_b <-  -0.03675   # parameter b in Taylor et al. 2012, Model 2.
par_g <- 0.3311  # parameter g in Taylor et al. 2012, Model 2

#Calculating and plotting probabilities
AttachmentDisp<-f_airflow(D = dist, a = par_a, b = par_b,g = par_g)
plot(AttachmentDisp~D,xlab="Distance (Km)", ylab="Probability", pch=16)


}
\references{
Taylor, K., Brummer, T., Taper, M. L., Wing, A., & Rew, L. J. (2012). Human‐mediated long‐distance dispersal: an empirical evaluation of seed dispersal by vehicles. Diversity and Distributions, 18(9), 942-951.
}
\author{
Maurizio Bagnara, Hanno Seebens
}
\keyword{CASPIAN}
