% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natural_kernel.R
\name{f_natural}
\alias{f_natural}
\title{Function for the terrestrial natural dispersal kernel.}
\usage{
f_natural(D, a, b)
}
\arguments{
\item{D}{numeric, a vector of distances between the points (in Km).}

\item{a}{numeric, shape parameter. Should be always >0}

\item{b}{numeric, shape parameter. Should be always >0. if b>1: thin-tailed kernel; if b<1: fat-tailed kernel;
Exponential if b=1, Gaussian if b=2}
}
\value{
numeric vector, including the probabilities of a species travelling the input distances by natural processes.
}
\description{
This function computes the probability of a species to reach a point in the network
starting from a different point by natural processes. Mathematically described as an exponential power law (see Nathan et al.2012).
}
\examples{
#generating vector of distances, in Km:
dist<-seq(from=0.1,to=10,length.out=100)

#assigning parameter values:
par_a <- 0.0001 # González-Martínez et al . 2006. Values for P. pinaster.
par_b <- 0.5 # González-Martínez et al . 2006. Values for P. pinaster.

#Calculating and plotting probabilities
natDisp<-f_natural(D = dist,a = par_a, b = par_b)
plot(natDisp~D,xlab="Distance (Km)", ylab="Probability", pch=16)


}
\references{
Gonzàlez-Martìnez, S. C., Burczyk, J., Nathan, R. A. N., Nanos, N., Gil, L., & Alia, R. (2006). Effective gene dispersal and female reproductive success in Mediterranean maritime pine (Pinus pinaster Aiton). Molecular Ecology, 15(14), 4577-4588.
Nathan, R., Klein, E. K., Robledo-Arnuncio, J. J., & Revilla, E. (2012). Dispersal kernels. In Dispersal ecology and evolution (pp. 187-210). Oxford: Oxford University Press.
}
\author{
Maurizio Bagnara, Hanno Seebens
}
\keyword{CASPIAN}
