% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConfigFile.R
\name{getConfigFile}
\alias{getConfigFile}
\title{Function to obtain the Configuration File for CASPIAN.}
\usage{
getConfigFile(outputDir, overwrite = FALSE)
}
\arguments{
\item{outputDir}{character, the path where the Configuration File will be saved}
}
\description{
This function allows the user to obtain the Configuration File needed to run CASPIAN. An existing file with the same name will be overwritten.
It is STRONGLY RECOMMENDED to change only the values of parameters and variables, not their names, or the model will not recognize them.
}
\examples{
#saving the configuration file in "C:"
getConfigFile("C:/")

#saving the configuration file in the current working directory, overwriting existing file:
getConfigFile(getwd(),overwrite=TRUE)

}
\author{
Maurizio Bagnara, Hanno Seebens
}
\keyword{CASPIAN}
